refactor - separating algorithms from UI
